Switch notification handling from tasks to instances
In order to support recurring task notifications propertly we switch from notifying tasks to notifying instances.
At present this might mean you get notifications for different instances of the same task. Not sure if that's a good or bad thing. In any case we're going to solve that in a future commit.